student_management/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.9
django-adminlte3==0.1.6
django-bootstrap-modal-forms==2.2.0
django-bootstrap3==21.1
django-ckeditor==6.2.0
django-crispy-forms==1.13.0
django-embed-video==1.4.0
django-forms-bootstrap==3.1.0
django-import-export==2.7.0
django-js-asset==1.2.2
django-widget-tweaks==1.4.9
et-xmlfile==1.1.0
idna==3.3
import-export==0.2.67.dev6
MarkupPy==1.14
mysql-connector-python==8.0.27
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.4.0
protobuf==3.19.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
sqlparse==0.4.2
tablib==3.1.0
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
